move css structure and add docstrings to js functions
