+++ frontend/src/styles/components/HelpModal.css
