+++ frontend/src/styles/components/MenuDropdown.css
