+++ frontend/src/styles/components/Statistics.css
